add ZipBasePath, refactor file
